requirements.txt
streamlit
yfinance
google-generativeai
plotly
python-dotenv
pandas
